rename to dashboardActions and fix connectivity issues
