+++ app/actions/dashboardActions.js
@@ -1,5 +1,5 @@
 // @flow
-import { api, rpc } from 'neon-js'
+import { api, rpc, RPCClient } from 'neon-js'
 import { extend } from 'lodash'
 import { createActions } from 'spunky'
 
@@ -13,14 +13,10 @@ type Props = {
   tokens: Array<TokenItemType>
 }
 
-export const ID = 'BALANCES_AND_BLOCKHEIGHT'
+export const ID = 'DASHBOARD'
 
-async function getBalances ({ net, address, tokens }: Props) {
-  const endpoint = await api.getRPCEndpointFrom({ net }, api.neoscan)
-
-  const client = new rpc.RPCClient(endpoint)
-  const blockHeight = await client.getBlockCount()
-
+async function getBalances (endpoint: string, { net, address, tokens }: Props) {
+  
   // token balances
   const promises = tokens.map(async token => {
     const { scriptHash } = token
@@ -59,15 +55,33 @@ async function getBalances ({ net, address, tokens }: Props) {
     })()
   )
 
-  return {
-    balances: extend({}, ...(await Promise.all(promises))),
-    blockHeight
-  }
+  return extend({}, ...(await Promise.all(promises)))
 }
 
 export default createActions(
   ID,
   ({ net, address, tokens }: Props = {}) => async (state: Object) => {
-    return getBalances({ net, address, tokens })
+    const endpoint = await api.getRPCEndpointFrom({ net }, api.neoscan)
+    const client = new rpc.RPCClient(endpoint)
+    
+    let balances = null
+    let blockHeight = null
+
+    try {
+      balances = await getBalances(endpoint, { net, address, tokens })
+    } catch (err) {
+      console.error(err)
+    }    
+
+    try {
+      blockHeight = await client.getBlockCount()
+    } catch (err) {
+      console.error(err)
+    }
+
+    return {
+      balances,
+      blockHeight
+    }
   }
 )
