+++ app/containers/App/Header/WalletBlockHeight/index.js
@@ -3,7 +3,7 @@ import { compose } from 'recompose'
 import { withData } from 'spunky'
 
 import WalletBlockHeight from './WalletBlockHeight'
-import balancesAndBlockHeightActions from '../../../../actions/balancesAndBlockHeightActions'
+import dashboardActions from '../../../../actions/dashboardActions'
 
 const mapBlockHeightDataToProps = (
   data
@@ -14,6 +14,6 @@ const mapBlockHeightDataToProps = (
     blockHeight: data.blockHeight
   }
 
-export default compose(withData(balancesAndBlockHeightActions, mapBlockHeightDataToProps))(
+export default compose(withData(dashboardActions, mapBlockHeightDataToProps))(
   WalletBlockHeight
 )
